# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Subcults API
  description: |
    Privacy-first platform for mapping underground music communities.
    Combines real-time location awareness, live audio streaming, and trust-based discovery.

    ## Authentication
    JWT-based authentication with two token types:
    - **Access tokens**: 15-minute expiry, includes `did` (Decentralized Identifier) claim
    - **Refresh tokens**: 7-day expiry, cookie-only

    Pass the access token via `Authorization: Bearer <token>` header or the `access_token` cookie.

    ## Privacy
    All public location data uses geohash-based jitter to prevent precise location tracking.
    Precise coordinates are only available when `allow_precise` consent is granted.

    ## Rate Limiting
    Per-endpoint rate limits are enforced. When exceeded, the API returns `429 Too Many Requests`
    with a `Retry-After` header.
  version: 0.1.0
  contact:
    name: Subcults
  license:
    name: MIT
servers:
  - url: /
    description: Current deployment
...
paths:
...
  /payments/onboard:
    post:
      operationId: onboardScene
      tags: [Payments]
      summary: Start Stripe Connect onboarding for a scene
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnboardSceneRequest'
      responses:
        '200':
          description: Onboarding URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnboardSceneResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Already onboarded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /payments/checkout:
    post:
      operationId: createCheckoutSession
      tags: [Payments]
      summary: Create a Stripe Checkout session
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutSessionRequest'
      responses:
        '200':
          description: Checkout session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutSessionResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /payments/status:
    get:
      operationId: getPaymentStatus
      tags: [Payments]
      summary: Get payment status
      security:
        - bearerAuth: []
      parameters:
        - name: session_id
          in: query
          schema:
            type: string
        - name: payment_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Payment status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentStatusResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Uploads ─────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT access token. Include as Authorization: Bearer <token> header or access_token cookie.'
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              enum:
                - validation_error
                - auth_failed
                - not_found
                - rate_limited
                - internal_error
                - forbidden
                - conflict
                - bad_request
                - invalid_palette
                - scene_deleted
                - scene_not_found
                - invalid_scene_name
                - duplicate_scene_name
                - invalid_time_range
                - missing_target
                - unsupported_type
                - invalid_weight
                - alliance_deleted
                - self_alliance
                - already_onboarded
                - not_onboarded
            message:
              type: string
              description: Human-readable error message

    # ── Geographic ──────────────────────────────────────────────────
...
    OnboardSceneRequest:
      type: object
      required: [scene_id]
      properties:
        scene_id:
          type: string
    OnboardSceneResponse:
      type: object
      properties:
        url:
          type: string
          description: Stripe onboarding link
        expires_at:
          type: string
          format: date-time
    CheckoutSessionRequest:
      type: object
      required: [scene_id, items, success_url, cancel_url]
      properties:
        scene_id:
          type: string
        event_id:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/CheckoutItemRequest'
        success_url:
          type: string
          format: uri
        cancel_url:
          type: string
          format: uri
    CheckoutItemRequest:
      type: object
      required: [price_id, quantity]
      properties:
        price_id:
          type: string
        quantity:
          type: integer
          format: int64
          minimum: 1
    CheckoutSessionResponse:
      type: object
      properties:
        session_url:
          type: string
        session_id:
          type: string
    PaymentStatusResponse:
      type: object
      properties:
        status:
          type: string
        amount_cents:
          type: integer
          format: int64
        fee_cents:
          type: integer
          format: int64
        currency:
          type: string
        updated_at:
          type: string
          format: date-time

    # ── Upload ──────────────────────────────────────────────────────
...